+++ crates/ryo3-core/src/lib.rs
@@ -1,8 +1,12 @@
 mod errors;
+mod map_pyerr;
 mod pymutex;
+mod pyparse;
 pub mod pystring;
 pub mod types;
 pub use pystring::{pystring_ascii_new, pystring_fast_new};
 
 pub use errors::FeatureNotEnabledError;
+pub use map_pyerr::map_py_value_err;
 pub use pymutex::{PyLock, PyMutex, map_poison_error};
+pub use pyparse::{PyFromStr, PyParse};