+++ crates/ryo3-std/src/net/socketaddr.rs
@@ -2,26 +2,27 @@ use crate::net::PyIpv6Addr;
 use crate::net::ipaddr::{IpAddrLike, PyIpAddr, PyIpv4Addr};
 use crate::net::ipaddr_props::IpAddrProps;
 use pyo3::prelude::*;
+use ryo3_core::{PyFromStr, PyParse};
 use std::hash::{Hash, Hasher};
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6};
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
-#[pyclass(name = "SocketAddrV4", frozen)]
+#[pyclass(name = "SocketAddrV4", frozen, immutable_type)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, PartialOrd, Ord)]
 pub struct PySocketAddrV4(pub(crate) SocketAddrV4);
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
-#[pyclass(name = "SocketAddrV6", frozen)]
+#[pyclass(name = "SocketAddrV6", frozen, immutable_type)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, PartialOrd, Ord)]
 pub struct PySocketAddrV6(pub(crate) SocketAddrV6);
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
-#[pyclass(name = "SocketAddr", frozen)]
+#[pyclass(name = "SocketAddr", frozen, immutable_type)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, PartialOrd, Ord)]
 pub struct PySocketAddr(pub(crate) SocketAddr);
@@ -84,9 +85,13 @@ impl PySocketAddrV4 {
     }
 
     #[staticmethod]
-    fn parse(s: &str) -> PyResult<Self> {
-        let sock = s.parse()?;
-        Ok(Self(sock))
+    fn from_str(s: &str) -> PyResult<Self> {
+        Self::py_from_str(s)
+    }
+
+    #[staticmethod]
+    fn parse(s: &Bound<'_, PyAny>) -> PyResult<Self> {
+        Self::py_parse(s)
     }
 
     #[classattr]
@@ -227,9 +232,13 @@ impl PySocketAddrV6 {
     }
 
     #[staticmethod]
-    fn parse(s: &str) -> PyResult<Self> {
-        let sock = s.parse()?;
-        Ok(Self(sock))
+    fn from_str(s: &str) -> PyResult<Self> {
+        Self::py_from_str(s)
+    }
+
+    #[staticmethod]
+    fn parse(s: &Bound<'_, PyAny>) -> PyResult<Self> {
+        Self::py_parse(s)
     }
 
     #[classattr]
@@ -367,9 +376,13 @@ impl PySocketAddr {
     }
 
     #[staticmethod]
-    fn parse(s: &str) -> PyResult<Self> {
-        let sock = s.parse()?;
-        Ok(Self(sock))
+    fn from_str(s: &str) -> PyResult<Self> {
+        Self::py_from_str(s)
+    }
+
+    #[staticmethod]
+    fn parse(s: &Bound<'_, PyAny>) -> PyResult<Self> {
+        Self::py_parse(s)
     }
 
     #[expect(clippy::wrong_self_convention)]